requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
scikit-learn
pandas
numpy